+++ RadioNetworks/NetworkParameters.cs
@@ -4,15 +4,15 @@ namespace RadioNetworks
     {
         #region Properties
 
-        public float ThroughputInMbps { get; set; }
-        public float BitErrorRate { get; set; }
-        public float BurstErrorRate { get; set; }
-        public float PacketLossPercentage { get; set; }
-        public float DelayInMsec { get; set; }
-        public float ResponseTimeInMsec { get; set; }
-        public float JitterInMsec { get; set; }
-        public float SecurityLevel { get; set; }
-        public float CostInUnitsPerByte { get; set; }
+        public double ThroughputInMbps { get; set; }
+        public double BitErrorRate { get; set; }
+        public double BurstErrorRate { get; set; }
+        public double PacketLossPercentage { get; set; }
+        public double DelayInMsec { get; set; }
+        public double ResponseTimeInMsec { get; set; }
+        public double JitterInMsec { get; set; }
+        public double SecurityLevel { get; set; }
+        public double CostInUnitsPerByte { get; set; }
 
         #endregion
     }
